Add files via upload
Changing passwords on Mac computers is a difficult thing to manange. This script attampts to fix any issues brought about by iCloud and keychain issues. Useful in Mac computer labs or other institutional Mac computers that are bound to AD, where the password may be changed on a different computer.